fix(comp): adds support to disable object in selector base. (#403)
**Pull Request Description**

The `SelectorBase `component now can be provided either a list strings
or a list o objects in the following structure:

{ text: "...", value: "...", disabled: "true/false" } 
(Format is based on the Bootstrap-Vue [Form Select Component's options
format](https://bootstrap-vue.org/docs/components/form-select))

If the `SelectorBase `component receives options as a list of strings:

Converts the array of strings to a lists of objects with the above
format to populate the `optionslist`.
Emits the `selectedOption `as a string.

If the `SelectorBase` component receives options as a list of objects
using the above format:

Populate the `optionslist `using the options directly.
Emit the `selectedOption `as a string.

To achieve this, `options` lists passed to the component through the
`options` prop are normalized into the `text`:, `value`:, `disabled`:
format. If the object is already in the expected format, it will not
update the list. If the object is not in the expected format, the
program has no checks to ensure that it is. The `Selected` has not
changed in functionality, it will return the `value`: of the
`optionsList` selected object. Finally, if an object is passed to
selector base with `disabled:` set to true, it will be greyed out and
wont be selectable.

The following tests were added to test the functionality:

1. `Prop is a list of strings and objects` tests the ability of selector
base to successfully create a options list from mixed types (objects and
strings). It also tests the reactivity when updating the prop.
2. `Props reflect can be disabled/enabled through flag` tests that props
are disabled and enabled as specified in the prop. It also tests that
the props are updated properly when the disabled flags are updated.

Closes #399 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: RolandLocke <roland.w.locke@gmail.com>
Co-authored-by: Grant Braught <braught@dickinson.edu>