feat(dev): added git hook restage for adding files  (#415)
**Pull Request Description**

Pursuant to #32, patched back changes from #412 

Closes #32 
---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: Ty Chermsirivatana <chermsit@dickinson.edu>
Co-authored-by: braughtg <braught@dickinson.edu>