fix(dev): installs vale via npm and configures it in setup.bash (#426)
**Pull Request Description**

The Vale linter had been installed in the `fd2dev` container. This
resulted in an issue when the version of vale installed became
incompatible with the vale styles that were available. For example, if a
newer version of vale became available that also required newer style
files. In that case when `vale sync` was run in `setup.bash` it would
fail because the vale styles were not compatible with the version of
vale in the container. This PR fixes this issue by installing vale via
`npm` in `setup.bash` and also runs `vale sync` in `setup.bash` ensuring
that we have the latest versions of each.

Closes #424 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.