feat(dev): Changes the development enviornment to use Codespaces instead of GitPod (#423)
**Pull Request Description**

Implements the launch of the linux-based development environment within
Codespaces. This used to rely on GitPod Classic which is being
decommissioned in April 2025. This PR adds a devcontainerr specification
that allows the linux-based development environment to be launched
within Codespaces and then accessed via a browser or via noVNC.

Closes #404 
Closes #418

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.