SelectorBaseRefactor (#427)
**Pull Request Description**

Refactored Selector Base. 

1. Cleaned redundant repeated code from the `selected()` watcher
2. Reassigned property and data watchers simplifying:
  - Properties being watched only update data in the component. 
- Separate watchers watch data within the component transforming when
necessary.

This is effectively accomplishes: Incoming property changes only trigger
a change in the data. Any further updates to the data should occur in
the component data not in the component's properties.
  
These changes hope to simplify the process of receiving and transforming
information within the component and reduce Vue's reactivity complexity.
4. Renamed a function to better reflect it's function

Provides an update to Issue #399 which has already been merged.
Closes #399

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: RolandLocke <roland.w.locke@gmail.com>