feat(comp): Adds `ActivePlantAssetPicklist` Component for Selecting Active Plantings by Location (#422)
**Pull Request Description**

This PR adds the `ActivePlantAssetPicklist` component, which allows
users to view and select active plantings in a given location. The
`ActivePlantAssetPicklist` component:

- Accepts a location name as a prop (e.g., `ALF`).  
- Displays all active plantings in the selected location.  
- Allows users to pick crops from the available plantings.  
- Passes relevant props to `PicklistBase`.  

Additionally:  
- Added tests for `ActivePlantAssetPicklist`.  
- Created an example page for `ActivePlantAssetPicklist`.  

Closes #419 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.