refactor(fd2): soil dist page for ActivePlantAsset component (#425)
**Pull Request Description**

This PR refactors the `soil_disturbance` entrypoint to use the
`ActivePlantAssetPicklist` component instead of `PicklistBase`.

### **Changes Made:**
- **Replaced `PicklistBase` with `ActivePlantAssetPicklist`** in the
`soil_disturbance` entrypoint.
- **Removed redundant code** from `soil_disturbance` that fetched and
formatted plant assets, as this logic is now handled by
`ActivePlantAssetPicklist`.
- Area calculation is now handled within `ActivePlantAssetPicklist`.
- The LocationSelector's `BedPicker` and `ActivePlantAssetPicklist`
visibility are now controlled via events emitted from
`ActivePlantAssetPicklist`

This refactor aligns with issue
**[#420](https://github.com/FarmData2/FarmData2/issues/420)** and
**depends on the completion of
[#419](https://github.com/FarmData2/FarmData2/issues/419)**.

**Closes #420**

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.