fix: Updates tests for sample DB with location J (#433)
**Pull Request Description**

Updates the test suite to match a new sample database that includes a
new location (field J) that contains no beds and no active plant assets.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.