feat: Creates a new fd2 menu icon (#432)
**Pull Request Description**

This pull request update the toolbar (gin) icon for the
FarmData2 modules.

closes #18 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: RolandLocke <roland.w.locke@gmail.com>