fix(lib): patches the permissions api call for deploy (#442)
**Pull Request Description**

Uses the `host` from the `document.url` to set the path to the
permissions api when the page making the call has been retrieved from a
farmOS instance.

Closes #441 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.