fix(dev): Fixes link to vale linter and makes setup.bash re-runnable (#443)
**Pull Request Description**

Ensures that vale linter is on the path and that it can be found by
VSCode.

In addition, setup.bash was cleaned up so that unnecessary tasks are not
performed if it is re-run in the fd2_dev container.

Closes #440

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.