Lint testing (#448)
**Pull Request Description**

Added eslint rules for v-on and v-bind to ensure longform is always used
instead of shortform notation.

Closes #387 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Co-authored-by: Niloy Saha
<143445417+niloy-saha-123@users.noreply.github.com>