fix(dev): uses sg to remove need to run fd2-up.bash multiple times (#416)
**Pull Request Description**

Modifies the `fd2-up.linux.bash` to use `sg` to rerun the `fd2-up.bash`
script when the user is added to a new group. Doing this instead of
using `newgrp` removes the need for the user to explicitly re-run
`fd2-up.bash`.

Closes #405

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Co-authored-by: Ty Chermsirivatana <28925758+WarpWing@users.noreply.github.com>