feat(dev): adds scripts and documentation for deploying a test instance (#450)
**Pull Request Description**

This PR provides scripts and directions for how to deploy a test
instance of FarmData2 to a Digital Ocean droplet. This will enable
FarmData2 to be accessed via mobile device from the wider internet and
will make it possible for workers on the Dickinson Farm to test its
functionality.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.