fix(dev): updates synchronization step (#451)
**Pull Request Description**

Updates the synchronization step of the workflow to use the "Sync Fork"
button in GitHub rather than pushing to origin. This is because
`development` is a protected branch and it is not possible to push
directly to it.

Closes #384 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.