fix(dev): updates specs for deploy (#452)
**Pull Request Description**

The test deploy can be run on a Droplet with lesser specs than
originally given. This costs half as much to run so the lesser specs are
worth using.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.