fix(comp): left aligns invalid feedback message (#456)
**Pull Request Description**

Modified the code in PicklistBase.vue file to shift the
invalidFeedbackText under the first column.

Closes #446 

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Co-authored-by: Niloy Saha <143445417+niloy-saha-123@users.noreply.github.com>