feat(fd2): creates launch pad for all FD2 features (#454)
**Pull Request Description**

I created buttons to direct to different pages like soil disturbance,
cover crop seeding etc. And I categorized the buttons to their respective
types. I wrote the tests to see if the page is working properly.

Closes #447 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Co-authored-by: Fej Esievo <116123894+fejesievo@users.noreply.github.com>
Co-authored-by: braughtg <braught@dickinson.edu>