feat(dev): updates addComponent.bash script and templates (#458)
**Pull Request Description**

Updates the `addComponent.bash` script so that in addition to creating
the new component it also creates an example page for it in the
`farm_fd2_examples` module. The templates for the new component were
updated to more closely match current idioms for components. Templates
were added for the example page.

The `buildCompExListPage.bash` script which builds the list of component
examples in the `App.vue` of the `component_examples` entry point in the
`farm_fd2_examples` module. Running this script ensures that the list is
in sync with the actual components that exist.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.