fix(fd2): adds farm_transplanting as dependency (#468)
**Pull Request Description**

Adds the farm_transplanting module as a dependency of the farm_fd2
module.

Related to #370 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.