feat(comp): creates an example page for TraySizeSelector (#466)
**Pull Request Description**

Created an example page for TraySizeSelector and also created a test
file to check if the page exists.

Closes #464 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.