fix(dev): runs pre-commit e2e tests on live server (#476)
**Pull Request Description**

Changes the lint staged configuration to run the e2e tests using the
live farmOS server instead of the dev server. General practice has
become to manually run tests against the live server. Also all non-e2e
tests (e.g. unit/component) use the live server for API calls, so the
live server must be running. So for consistency, the e2e tests now run
on the live server as well.

Closes #462

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.