fix(dev): updates addEntryPoint.bash testing and script output (#477)
**Pull Request Description**

Updates the `addEntryPoint.bash` script to run e2e tests on the live
server and also run the `lib.js` unit tests. The output of the script is
also cleaned up to better match `addComponent.bash`.

Closes #453

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.