fix(dev): updates addComponent.bash error instructions (#478)
**Pull Request Description**

Updates the instructions for recovering from an error in the
`addComponent.bash` script to give `git` commands instead of just
describing the operations to be performed.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.