fix(dev): adds linting rules for .only and console.log (#469)
**Pull Request Description**

Added linting rules to prohibit console.log and .only statements in test files.

Closes #19 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.