fix(fd2): removes unnecessary quantity deletes (#481)
**Pull Request Description**

Added if conditionals to the lib file in cover crop seeding and tray
seeding files, and added more units tests in each of their submitError
test files.

Closes #363 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: braughtg <braught@dickinson.edu>