docs(lib): adds @example JSDoc tag to runTransaction function (#487)
Closes #221 

Enhanced the documentation for the
[runTransaction](vscode-file://vscode-app/usr/share/code/resources/app/out/vs/code/electron-sandbox/workbench/workbench.html)
function by adding an @example tag to demonstrate its usage with sample
operations.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: braughtg <braught@dickinson.edu>