feat(fd2): adds harvest units to farmOS plant types (#484)
**Pull Request Description**

This PR adds the `fd2_harvest_unit` and `fd2_unit_conversions`
attributes to the farmOS `plant_type` vocabulary so that all
`plant_type` terms (i.e. all crops) will have a default unit of harvest,
and a list of additional units in which it can be harvested. The
additional units will all have a conversion factor to convert them into
the default unit.

In addition, this PR also updates documentation to describe how to add
additional attributes to vocabularies.

Closes #472

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.