feat(dev): install correct db version when branch changes (#507)
**Pull Request Description**

Stashes the currently installed db version and artifact in
`.fd2dev/db.conf` when `installDB.bash` is run. This value is then used
by the `post-checkout` Git hook to change the db version when switching
to a branch. This ensures that the db installed is compatible with the
code in the branch.

Closes #211

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.