fix(comp): patches selector in crop related component popup tests  (#521)
**Pull Request Description**

Patches typo in the popup tests for the harvest unit field in the
CropSelector and MultCropSelector components. It also adds component e2e
tests to the lintstaged configuration.

Closes #520 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.