fix(dev): reorder devcontainer tasks to run heartbeat earlier (#525)
**Pull Request Description**

Heartbeat listener was not starting if the browser tab showing the
"Running" screen was never active. This would be a common issue as once
the dev env is started there is no reason to make the tab with the
"Running" screen active again. The heartbeat listener is now started
before waiting to display the "Running" Screen which should address the
issue.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.