fix(dev): patches codespace dev container run heartbeat listener in background (#526)
**Pull Request Description**

Patches the .devcontainer.json so that the heartbeat server runs in the
background. With the prior commit the heartbeat server process (ncat)
was being terminated when the task completed.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: Grant Braught <braught@dickinson.edu>