fix(dev): patch to the devcontainer to run tasks in separate terminals (#527)
Run the task that switches to the "Running" screen and the task that
runs the heartbeat listener in separate terminals. This allows them both
to complete.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.