fix(dev): reveals terminal for the heartbeat listener (#530)
**Pull Request Description**

The terminal in which the heartbeat listener is running in the codespace
is revealed to ensure that text displayed in that terminal counts as
interaction and resets the timeout timer.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.