fix(dev): patch installDB and compose for proper .fd2 location (#531)
**Pull Request Description**

Fixes the Docker `compose.yml` and `installDB.bash` scripts to use the
up to date location for `.fd2` directory within the FarmData2
repository.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.