patches installDB.bash to use info in .fd2dev/db.conf by default (#532)
**Pull Request Description**

The default behavior of `installDB.bash` had been to install the latest
database release. This patch changes that so that the default behavior
is to install the version of the database that is recorded in the
`.fd2dev/db.conf` file.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.