updates to latest database version with harvest units (#533)
**Pull Request Description**

Updates the `.fd2dev/db.conf` to point to version 3.7.0 of the database
that includes the harvest units and unit conversions for plant type
vocabulary terms.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.