fix(test): patches units lib tests for added units (#535)
**Pull Request Description**

The updated sample database added additional units to the Units
taxonomy. This reordered the units and changed the number of units,
which broke some tests. This PR patches those tests to match the
database.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.