fix(docs): updates to install docs (#536)
**Pull Request Description**

Updates the install documentation by:
- Removing the GitPod install link in the documentation index page.
- Adding more information to the instructions in the Codespaces install
about configuring Git and authenticating with GitHub.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.