03 vue1 application starter (#33)
**Pull Request Description**

A working solution to the 02-HTML-CSS-Application assignment that serves
as Starter code for the 03-Vue1-Application assignment.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.