Adds crop as a query parameter to the fd2_plant_assets API endpoint (#99)
**Pull Request Description**

Adds crop as a query parameter to the fd2_plant_asset api so that it is
possible to query for all of the plants of a particular type of crop.
The getPlantAssets function in the farmOSUtil library is also updated to
allow fetching all plants of a given crop. All ripple effects of this
change have also been patched.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.