fix(fd2): adds harvest log category (#546)
**Pull Request Description**

Adds a log category for harvests. The code to add the category is in the
module's `.install` file.

Also added is a script that will uninstall and reinstall the `farm_fd2`
module. This is useful when working on the module files. It is necessary
because `farm_fd2` adds custom fields and drupal will not allow it to
uninstalled while those fields exist. So this script deletes those
fields, uninstalls `farm_fd2`, builds `farm_fd2`, and then reinstalls
`farm_fd2`.


---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.