fix(fd2): adds `harvest` log category (#548)
**Pull Request Description**

- Adds a `harvest` category to the "log categories" taxonomy.
- Adds a script to reinstall the `farm_fd2` module.
- Needed because `farm_fd2` adds custom fields and drupal will not allow
it to be uninstalled because data would be deleted.
- The script deletes the custom fields, uninstalls the module, rebuilds
the module, and then reinstalls the module.
- Patches the tests for the log categories functions in the `farmosUtil`
library.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.