adds harvest log category and associated patches (#121)
**Pull Request Description**

Adds a harvest log category to the log category taxonomy. Patches the
associated tests and updates to the version of the sample database that
contains this log category.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.