sets log category in created harvest logs (#125)
**Pull Request Description**

Updates the `farmosUtil.harvest.js` library and associated tests so that
`harvest` is included in the log categories of the created log.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.