patches container names in reinstallFD2Module.bash (#161)
**Pull Request Description**

The reinstall script appended `_school` to the container names to avoid
a name clash on my machine and these slipped into the original PR by
mistake.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.