Improve docs on components used in 09 (#186)
**Pull Request Description**

Makes the docs for CommentBox, CropSelector, DateSelector, and
NumericInput more complete and easier to understand.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.