Patch the unitList in the 09 tutorials. (#187)
**Pull Request Description**

Initializes the unitList to [] instead of a static list. Doesn't change
the overall behavior, but is a cleaner solution.



**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.