09 fd3 activity starter (#211)
**Pull Request Description**

Starter code for 09 Activity, contains solution to 09 Tutorials + In
class live-coded update to use v-model.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.