fix(dev): fix typos in homepage_template.md (#555)
Fix typos

**Pull Request Description**
Fixes 4 visible typos and one in comments

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: atap <55597131+aarontapper@users.noreply.github.com>