Typo fixes on homepage_template.md (#221)
Fixes four typos on this page that have been bugging me for a while

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: atap <55597131+aarontapper@users.noreply.github.com>