add bugInfo script and update PR template (#229)
**Pull Request Description**

Updates the PR template to be more explicit about what is expected.

Adds the `bugInfo.script` that is used with the bug reporting issue
template added in #228.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.