Update bugInfo script to check for local changes (#230)
### Purpose

The `bugInfo.bash` script should check for local changes and only
continue to the output if no local changes have been made to the
`development` branch.

### Verification steps

1. Make a local change.
2. Run `bugInfo.bash`.
   - It should report that there are local changes and exit.
3. `git stash` or `git restore` the local changes.
4. Run `bugInfo.bash` again.
   - It should now run through and generate its output.

### Approach

The `git status --short` command was used to check for any local
changes.

### Related issues

N/A

### Additional information

N/A

### Licensing Certification

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: Grant Braught <braught@dickinson.edu>