updates formatting of PR template and adds testing section (#231)
### Purpose

To update the formatting of the PR template by removing the space
between the heading and the italicized instructions. Also touched up a
bit of phrasing.

### Verification steps

- Create a new PR and verify the italicized content appears directly
below the heading.

### Approach

N/A

### Related issues

N/A

### Additional information

N/A

### Licensing Certification

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.