Update devcontainer.json to pin base image to bookworm (#556)
A new Debian base image was released that was not compatible with the
Docker-In-Docker features. This patch pegs the base image being used to
be compatible with the version of the Docker-in-Docker feature that is
installed.

**Pull Request Description**

_Describe what your Pull Request does here_

_Include a "Closes #" line for the issue your PR closes._

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: Grant Braught <braught@dickinson.edu>