Codespace fix (#250)
### Purpose

GitHub codespaces appears to have changed a security setting requiring
that terminals be visible when running a VSCode task. This PR changes
the tasks in `.devcontainer` so that they run in a visible terminal.
This is less the aesthetic from the UI perspective but it does work as a
patch for now,

### Licensing Certification

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.