fix(dev): .devcontainer patch (#558)
Patches the .devcontainer to reveal the terminal in which VSCode tasks
are running. This appears to be a change in how GitHub codespaces
execute the VSCode tasks. If the terminal was not revealed the first
task would fail and no further tasks would run. Revealing the terminal
window solved this issue.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: Grant Braught <braught@dickinson.edu>