Update README.md (#287)
### Purpose
_A clear and concise description of the purpose of this pull request._

### Verification steps
_A clear and concise set of steps that will enable the reviewer to
manually verify that this pull request accomplishes its purpose._

### Approach
_A clear and concise description of the approach taken with rationale
for design decisions that were made._

### Testing
_A clear and concise description of automated tests that have been
modified or added to verify changes that were made._

### Related issues
_Include "Relates to #" and/or "Closes #" lines to link to issues
connected to this pull request._

### Additional information
_Include any additional information you think might be helpful here._

### Licensing Certification

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: Grant Braught <braught@dickinson.edu>