feat(dev): updates pr template and adds bug/feature templates (#568)
**Pull Request Description**

Updates the pull request template and adds separate templates for
feature requests and bug reports.

Closes #229 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.