feat(dev): adds the bugInfo.bash script (#570)
### Purpose

Adds the `bugInfo.bash` script. This script is run to generate some
useful information about the current state of the `development` branch
when a bug report is being filed. The Bug Report template now requests
that output of this script as part of the bug report.

### Licensing Certification

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **the contributor (and any listed co-authors) certify
that they meet the terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.