fix(dev): Adds execute permission to the `bugInfo.bash` script (#571)
### Purpose

Adds the execute permission to the script.

### Licensing Certification

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **the contributor (and any listed co-authors) certify
that they meet the terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.